#25 rename: TitleLog 를 title 패키지 하위로 이동
