+++ src/main/java/com/idea5/four_cut_photos_map/domain/member/service/MemberService.java
@@ -7,7 +7,7 @@ import com.idea5.four_cut_photos_map.domain.member.entity.Member;
 import com.idea5.four_cut_photos_map.domain.member.repository.MemberRepository;
 import com.idea5.four_cut_photos_map.domain.title.entity.Title;
 import com.idea5.four_cut_photos_map.domain.title.service.TitleService;
-import com.idea5.four_cut_photos_map.domain.titleLog.entity.TitleLog;
+import com.idea5.four_cut_photos_map.domain.title.entity.TitleLog;
 import com.idea5.four_cut_photos_map.global.common.RedisDao;
 import com.idea5.four_cut_photos_map.security.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
