+++ src/main/java/com/idea5/four_cut_photos_map/domain/title/entity/TitleLog.java
@@ -1,8 +1,7 @@
-package com.idea5.four_cut_photos_map.domain.titleLog.entity;
+package com.idea5.four_cut_photos_map.domain.title.entity;
 
-import com.idea5.four_cut_photos_map.domain.title.entity.Title;
-import com.idea5.four_cut_photos_map.global.base.entity.BaseEntity;
 import com.idea5.four_cut_photos_map.domain.member.entity.Member;
+import com.idea5.four_cut_photos_map.global.base.entity.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
